requirements.txt
pymupdf==1.21.1
python-dateutil==2.8.1
dateparser==1.1.6
requests==2.25.1
python-decouple==3.7
regex==2023.5.5
flake8==6.0.0
pre-commit==3.3.3
filetype==1.2.0
deepdiff==6.3.0
fastapi==0.99.1
jinja2==3.1.2
python-multipart==0.0.6
uvicorn[standard]==0.22.0
markdown==3.4.3
secure==0.3.0
mysql-connector-python==8.1.0
